# Removed allocation from `filter_pixel()`