Achieved via always computing with 4 channels using an array of length 4 since const generic expressions are unstable.