@ripytide 2024-05-27
Okay now that #654 was merged this comparison becomes simpler.
In the generic case (not exactly 3x3) this version reduces the code duplication and is simply faster.
![image](https://github.com/image-rs/imageproc/assets/62516857/f747e182-8ef4-4ff5-bc3f-7063de842ce1)