@ripytide 2024-05-27
@cospectrum can you share your implementation of `filter3x3()` as well so I can try to reproduce your results?