@cospectrum 2024-05-27
> @cospectrum can you share your implementation of `filter3x3()` as well so I can try to reproduce your results or are you using the `filter3x3_reference()`?

It's not "my" implementation. It's from 0.25