@ripytide 2024-05-27
Now I suppose I just need to track down which specific build setting is making such a big difference.