@ripytide 2024-05-27
When using `debug=false` now the two implementation are both `3,300,000` but `master_025` is still `2,000,000` which doesn't make any sense at all?