@cospectrum 2024-05-28
If I bench `0.25` against `master`, `master` will have the same settings as your branch, but `master` has the same speed as `0.25`, while yours will be slower