@theotherphil 2024-05-28
I'll try this PR vs the current master on my MacBook when I have time within the next few days.