@ripytide 2024-05-28
I think the setting which changes which implementation is faster is `codegen-units`:

Here I've benchmarked the default compile settings vs `codegen-units=1` which is used by the benchmarks normally.
```toml
[profile.release]
codegen-units = 1
```
## Results
```
test master_codegen=1 ... bench:   2,375,138 ns/iter (+/- 3,180)
test pr_codegen=1 ... bench:   1,974,065 ns/iter (+/- 6,905)
test master_codegen=16 ... bench:   2,489,427 ns/iter (+/- 10,247)
test pr_codegen=16 ... bench:   4,585,778 ns/iter (+/- 27,787)
``