@theotherphil 2024-06-07
I’ve not had much time in the last couple of weeks but I still intend to benchmark this branch against master on my MacBook.